Added @since to hasNext and hasNotNext functions and refactored tests a bit
