+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/test/kotlin/ch/tutteli/atrium/api/fluent/en_GB/IterableAssertionsSpec.kt
@@ -3,6 +3,7 @@ package ch.tutteli.atrium.api.fluent.en_GB
 import ch.tutteli.atrium.creating.Expect
 import ch.tutteli.atrium.specs.fun0
 
-object IterableHasNotNextAssertionsSpec : ch.tutteli.atrium.specs.integration.IterableHasNotNextAssertionsSpec(
+object IterableAssertionsSpec : ch.tutteli.atrium.specs.integration.IterableAssertionsSpec(
+    fun0(Expect<Iterable<Int>>::hasNext),
     fun0(Expect<Iterable<Int>>::hasNotNext)
 )
