--- apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/test/kotlin/ch/tutteli/atrium/api/fluent/en_GB/IterableHasNextAssertionsSpec.kt
@@ -1,8 +1,0 @@
-package ch.tutteli.atrium.api.fluent.en_GB
-
-import ch.tutteli.atrium.creating.Expect
-import ch.tutteli.atrium.specs.fun0
-
-object IterableHasNextAssertionsSpec : ch.tutteli.atrium.specs.integration.IterableHasNextAssertionsSpec(
-    fun0(Expect<Iterable<Int>>::hasNext)
-)
